[added] absoluteTime param
when set, relative times for from and until will be converted to
absolute time. when generating links it can be useful to have absolute
times to retain the data you are attempting to display, but annoying to
have to build up the from and until string.

configs now have an absoluteTime option (defaults to false). to enable
this behavior, set it to true.